+++ src/router/router.tsx
@@ -5,7 +5,7 @@ import GomaoPage from "@pages/Gomao/Gomao";
 import AdminLogin from "@pages/Admin/AdminLogin";
 import AdminLayout from "@pages/Admin/AdminLayout";
 import AdminDashboard from "@pages/Admin/AdminDashboard";
-import NewMPUploader from "@pages/Admin/NewMPUploader";
+import NewMPUploader from "@pages/Admin/UploadMP";
 
 export const routers: RemixRouter = createBrowserRouter([
   {
